Reorganized and cleaned up: data folder
