+++ src/components/Contact.jsx
@@ -1,32 +1,5 @@
-import { FiMail } from "react-icons/fi";
-import { FaLinkedin, FaInstagram, FaGithub } from "react-icons/fa";
-
-const contacts = [
-  {
-    icon: <FiMail />,
-    label: "annabrekke98@gmail.com",
-    href: "mailto:annabrekke98@gmail.com",
-    className: "text-link",
-  },
-  {
-    icon: <FaLinkedin />,
-    label: "LinkedIn",
-    href: "https://www.linkedin.com/in/anna-aasprong-brekke-a571132b0/",
-    className: "text-link",
-  },
-  {
-    icon: <FaInstagram />,
-    label: "Instagram",
-    href: "https://www.instagram.com/annabrekke/",
-    className: "text-pink-600",
-  },
-  {
-    icon: <FaGithub />,
-    label: "GitHub",
-    href: "https://github.com/AnnaAaBrekke",
-    className: "text-text",
-  },
-];
+import { contacts } from "../data/contacts";
+import { IconMap } from "../utils/iconMap";
 
 const Contact = () => {
   return (
@@ -44,7 +17,7 @@ const Contact = () => {
             rel="noopener noreferrer"
             className={`flex items-center gap-2 ${className} hover:underline transition`}
           >
-            {icon}
+            {IconMap[icon]}
             {label}
           </a>
         ))}
