+++ src/components/Timeline.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { educationEntries, workEntries } from "../data/timeline";
 
 const TimelineSection = ({ title, entries }) => (
   <div className="mb-16">
@@ -15,52 +16,11 @@ const TimelineSection = ({ title, entries }) => (
   </div>
 );
 
-const Timeline = () => {
-  const education = [
-    {
-      institution: "Noroff College",
-      description: "Frontend Developer | Aug 2023 - Jun 2025",
-    },
-    {
-      institution: "Universitetet i Sørøst-Norge",
-      description:
-        "Master's Degree in Primary School Education | Aug 2018 - Jun 2023",
-    },
-    {
-      institution: "Drammen Toppidrett VGs",
-      description: "High School | Aug 2015 - Jun 2018",
-    },
-  ];
-
-  const workExperience = [
-    {
-      institution: "Lier Videregående skole",
-      description: "Teacher in High School | Jan 2022 - Present",
-    },
-    {
-      institution: "Åskollen Barneskole",
-      description: "Teaching Assistant | Jan 2017 – Jun 2022",
-    },
-    {
-      institution: "Drammen kommune: Sommerskole",
-      description: "Activity Leader | Jun 2019 - Aug 2021",
-    },
-    {
-      institution: "Digitale Læringsverkstedet",
-      description: "Student Assistant | Aug 2019 - Nov 2019",
-    },
-    {
-      institution: "Torshov Sport",
-      description: "Sales Associate | Mar 2016 - Aug 2018",
-    },
-  ];
-
-  return (
-    <section id="timeline" className="p-8">
-      <TimelineSection title="Education" entries={education} />
-      <TimelineSection title="Work Experience" entries={workExperience} />
-    </section>
-  );
-};
+const Timeline = () => (
+  <section id="timeline" className="p-8">
+    <TimelineSection title="Education" entries={educationEntries} />
+    <TimelineSection title="Work Experience" entries={workEntries} />
+  </section>
+);
 
 export default Timeline;
