+++ src/data/contacts.js
@@ -1,0 +1,26 @@
+export const contacts = [
+  {
+    icon: "FiMail",
+    label: "annabrekke98@gmail.com",
+    href: "mailto:annabrekke98@gmail.com",
+    className: "text-link",
+  },
+  {
+    icon: "FaLinkedin",
+    label: "LinkedIn",
+    href: "https://www.linkedin.com/in/anna-aasprong-brekke-a571132b0/",
+    className: "text-link",
+  },
+  {
+    icon: "FaInstagram",
+    label: "Instagram",
+    href: "https://www.instagram.com/annabrekke/",
+    className: "text-pink-600",
+  },
+  {
+    icon: "FaGithub",
+    label: "GitHub",
+    href: "https://github.com/AnnaAaBrekke",
+    className: "text-text",
+  },
+];
