+++ src/data/introCards.js
@@ -1,0 +1,31 @@
+export const introCards = [
+  {
+    id: "frontend",
+    title: "Frontend Developer",
+    description:
+      "Graduated from Noroff College with a degree in Frontend Development.",
+  },
+  {
+    id: "education",
+    title: "Master’s in Education",
+    description:
+      "Graduated from USN. Master’s thesis focused on algorithm awareness in Social Studies.",
+  },
+  {
+    id: "handball",
+    title: "Handball Captain",
+    description:
+      "Captain at Nordstrand Damer Elite, competing in Norway’s first division.",
+  },
+  {
+    id: "teacher",
+    title: "High School Teacher",
+    description: "Teaching at Lier Videregående Skole since 2022.",
+  },
+  {
+    id: "media",
+    title: "Media Creator",
+    description:
+      "Passionate about photography, video editing, and AI-generated images.",
+  },
+];
