+++ src/data/timeline.js
@@ -1,0 +1,38 @@
+export const educationEntries = [
+  {
+    institution: "Noroff College",
+    description: "Frontend Developer | Aug 2023 - Jun 2025",
+  },
+  {
+    institution: "Universitetet i Sørøst-Norge",
+    description:
+      "Master's Degree in Primary School Education | Aug 2018 - Jun 2023",
+  },
+  {
+    institution: "Drammen Toppidrett VGs",
+    description: "High School | Aug 2015 - Jun 2018",
+  },
+];
+
+export const workEntries = [
+  {
+    institution: "Lier Videregående skole",
+    description: "Teacher in High School | Jan 2022 - Present",
+  },
+  {
+    institution: "Åskollen Barneskole",
+    description: "Teaching Assistant | Jan 2017 – Jun 2022",
+  },
+  {
+    institution: "Drammen kommune: Sommerskole",
+    description: "Activity Leader | Jun 2019 - Aug 2021",
+  },
+  {
+    institution: "Digitale Læringsverkstedet",
+    description: "Student Assistant | Aug 2019 - Nov 2019",
+  },
+  {
+    institution: "Torshov Sport",
+    description: "Sales Associate | Mar 2016 - Aug 2018",
+  },
+];
