+++ src/utils/IconMap.jsx
@@ -1,0 +1,10 @@
+import { FiMail, FiGlobe } from "react-icons/fi";
+import { FaLinkedin, FaInstagram, FaGithub } from "react-icons/fa";
+
+export const IconMap = {
+  FiMail: <FiMail />,
+  FiGlobe: <FiGlobe />,
+  FaLinkedin: <FaLinkedin />,
+  FaInstagram: <FaInstagram />,
+  FaGithub: <FaGithub />,
+};
